pyproject.toml
[project]
name = "write-up-agent-general"
version = "0.1.0"
description = "Assist by writing content for social profiles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cryptography>=44.0.0",
  "fastapi[standard]>=0.115.6",
  "google-ai-haystack>=4.0.1",
  "google-generativeai>=0.8.3",
  "haystack-ai>=2.8.1",
  "python-dotenv>=1.0.1",
  "streamlit-cookies-controller>=0.0.4",
  "streamlit>=1.41.1",
  "typing>=3.10.0.0",
]
